Use glxinfo instead of lspci for querying GPU
By using glxinfo, the exact model of the GPU in use can be queried
instead of just a generic family of GPUs, which lspci provides. For
example, in the event that someone had an AMD Radeon RX 5700 XT, lspci
would say that the current GPU could be anything from a 5600 OEM model
all the way up to a 5700 XT, which is not necessarily correct. Using
glxinfo will return the full name of the GPU in use for OpenGL
applications and will also specify information such as the manufacturer
call sign and the LLVM version mesa was built against.